Use go modules to freeze dependencies
As raised in #69 the use of `go get -t ./...` is systematically downloading the latest version of files.
Lately, go exposes a new way to handle dependencies: modules. Those are freezing version of libraries to ensure reproducible builds.
This means there will hence be the need to update versions of dependencies and bump them. Though this would be a conscious update, that can be tested prior to merge.

For users who clone the repository in `$GOPATH/src/github.com/whalebrew/whalebrew` this means exposing the environment variable `GO111MODULE=on`