Add release workflow
Add an automation to systematically maintain a release up to date when a
new tag is pushed.
This workflow supports being restarted at any step as it reconciles
existing release for the given tag and forces synchronising existing
release assets as if the release was made from scratch.

Now, release are as simple as pushing a new tag.

Still missing: generating automatically the release change log and
opening the new version in-code