Feat: add support to search on any docker registry
Refactor how search is actually working to allow hooking and injecting
several search methods.
The alternative search method is using the registry
[catalog](https://docs.docker.com/registry/spec/api/#catalog) listing
all images in the registry and filtering them according to the searched
term.

Though this method is particularly inefficient with large registries.
We should at some point implement some kind of cache to speed up
consequent searches and avoid over-loading the remote server.
Similarly, to improve performances, we should probably think of a way to
integrate directly with the registries APIs so we can get better
performances.

Finally, this method can not be used on docker hub, that still uses the
dedicated docker hub search method, as the catalog has been
[disabled on docker hub](https://stackoverflow.com/a/37518393)
because it would list all the images available on docker hub

fixes: #102