Support images built with buildkit
buildkit built images are not providing the ContainerConfig we are
reading labels from. Instead, everything is populated in the Config field
of `inspect`.

We should eventually have better visibility about those fields in this
issue: https://github.com/docker/cli/issues/2868

Meanwhile, we should support it the best way possible.

Benefit from having to do this change to mutualise the label parsing
logic so we are consistent in how we manage the label lookup and
encoding (prefer yaml and fallback to string when possible)

Fixes: #119