Detect package or image name in uninstall
While the install commands accepts an image name, the uninstall command
expects the package name.
This is confusing for users as the command is not symetrical.

Implement a simple detection mechanism that allows to detect the
relevant package to be deleted so the uninstall command can be used
using the image name as the install endpoint accepts. The previous
behaviour to be able to uninstall a package from its name is kept

fixes #162

Change-Id: Ibdddfdb9da2680e6c0c2866dc15e27d331c20f47