Add support for podman
Context
---

We currently have more and more ways to run containers on computers.

The most obvious one is podman which provides an cli interface very similar
to the docker one, and mostly compatible, but with a different executable.

Problem
---

Even though users aliases and remaps docker to use podman underneath,
whalebrew uses directly the raw docker socket at the time of installing
and linting charts.

As podman does not expose any docker socket, this makes it impossible to
install or lint charts when podman is used as a runtime.

Goals
---

Enable the use of podman

Remove the requirement of the raw docker socket

Change-Id: I10aa352bd31263fa67a375f969fabf681c87243e

Fixes: #200
Fixes: #228