Add support for file-based configuration (#251)
* Add support for file-based configuration

As raised in #191, currently, the only way to customize the install path
is to use environment variables.

Provide more flexibility to configure it in different ways.

This implementation searches for the current whalebrew config path
and falls-back following the specifications of the [Free desktop XDG]
environment variables to look up for the relevant config file.

In case an environment variable has been provided, it will automatically
overwrite the matching content from the config file.

There is no surcharge mechanism to extend in the User home, the global
configuration files.

[Free desktop XDG]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

* Update README.md

* Update README.md